refactor : 스타일드 컴포넌트 구조 변경 [#19]
